backend/requirements.txt
django>=4.2.0
django-ninja>=1.0.1
django-oauth-toolkit>=2.3.0
pyjwt>=2.8.0
pydantic>=2.6.0
email-validator>=2.1.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.5 